requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
requests
google-api-python-client